+++ lib/components/ScenesManager/ScenesManager.tsx
@@ -8,7 +8,7 @@ import {
   migrateScene,
   ScenesContext,
 } from "../../contexts/SceneContext/ScenesContext";
-import { FariEntity } from "../../domains/FariEntity/FariEntity";
+import { FariEntity } from "../../domains/fari-entity/FariEntity";
 import { Manager } from "../Manager/Manager";
 
 type IProps = {};